modularize test suite
