+++ test/testsets/doctests.jl
@@ -1,0 +1,3 @@
+@testset "doctests" begin
+    doctest(Drifters; manual = false)
+end
